Satisfy sidebar CI policy checks
