+++ Packages/macOS/CmuxGit/Sources/CmuxGit/Probe/CIStatusGraphQL/GitHubGraphQLDynamicCodingKey.swift
@@ -1,0 +1,14 @@
+struct GitHubGraphQLDynamicCodingKey: CodingKey {
+    let stringValue: String
+    let intValue: Int?
+
+    init(stringValue: String) {
+        self.stringValue = stringValue
+        self.intValue = nil
+    }
+
+    init?(intValue: Int) {
+        self.stringValue = String(intValue)
+        self.intValue = intValue
+    }
+}
