+++ Packages/macOS/CmuxGit/Sources/CmuxGit/Probe/CIStatusGraphQL/GitHubGraphQLRequest.swift
@@ -1,0 +1,4 @@
+struct GitHubGraphQLRequest: Encodable {
+    let query: String
+    let variables: [String: String]
+}
