+++ Packages/macOS/CmuxGit/Sources/CmuxGit/Probe/CIStatusGraphQL/GitHubPullRequestCIStatusGraphQLCommit.swift
@@ -1,0 +1,3 @@
+struct GitHubPullRequestCIStatusGraphQLCommit: Decodable, Sendable {
+    let statusCheckRollup: GitHubPullRequestCIStatusGraphQLStatusCheckRollup?
+}
