+++ Packages/macOS/CmuxGit/Sources/CmuxGit/Probe/CIStatusGraphQL/GitHubPullRequestCIStatusGraphQLCommitNode.swift
@@ -1,0 +1,3 @@
+struct GitHubPullRequestCIStatusGraphQLCommitNode: Decodable, Sendable {
+    let commit: GitHubPullRequestCIStatusGraphQLCommit
+}
