+++ Packages/macOS/CmuxGit/Sources/CmuxGit/Probe/CIStatusGraphQL/GitHubPullRequestCIStatusGraphQLData.swift
@@ -1,0 +1,3 @@
+struct GitHubPullRequestCIStatusGraphQLData: Decodable, Sendable {
+    let repository: GitHubPullRequestCIStatusGraphQLRepository?
+}
