+++ Packages/macOS/CmuxGit/Sources/CmuxGit/Probe/CIStatusGraphQL/GitHubPullRequestCIStatusGraphQLError.swift
@@ -1,0 +1,3 @@
+struct GitHubPullRequestCIStatusGraphQLError: Decodable, Sendable {
+    let message: String
+}
