+++ Packages/macOS/CmuxGit/Sources/CmuxGit/Probe/CIStatusGraphQL/GitHubPullRequestCIStatusGraphQLPullRequest.swift
@@ -1,0 +1,4 @@
+struct GitHubPullRequestCIStatusGraphQLPullRequest: Decodable, Sendable {
+    let number: Int
+    let commits: GitHubPullRequestCIStatusGraphQLCommitConnection?
+}
