+++ Packages/macOS/CmuxGit/Sources/CmuxGit/Probe/CIStatusGraphQL/GitHubPullRequestCIStatusGraphQLStatusCheckRollup.swift
@@ -1,0 +1,3 @@
+struct GitHubPullRequestCIStatusGraphQLStatusCheckRollup: Decodable, Sendable {
+    let state: String?
+}
