+++ Packages/macOS/CmuxSettings/Sources/CmuxSettings/Keys/SidebarCatalogSection.swift
@@ -62,6 +62,7 @@ public struct SidebarCatalogSection: SettingCatalogSection {
         userDefaultsKey: "sidebarShowPullRequest"
     )
 
+    /// Whether sidebar pull-request badges include the latest CI/check rollup state.
     public let showPullRequestCIStatus = DefaultsKey<Bool>(
         id: "sidebar.showPullRequestCIStatus",
         defaultValue: false,
